Transfer common methods to ServiceManager
Add ServiceManager to transfer common methods
